The shared sight answer ages on 16 bits: subtract on 16 too
doomLevelTime is an int and the filed tic a short, so the difference could cross 65536
and read as fresh.  Both sides are narrowed before the subtraction, which is then the
modular one the test wants.

+++ game/doom/DOOM_VERBS.C
@@ -142,7 +142,8 @@ static int doomSee(Sprite *a,Sprite *b)
  if (k>=0)
     {m=sightMemo+(((a->s<<2)+k)&(DOOM_SIGHT_MEMO-1));
      if (m->k==(signed char)k && m->look==(short)a->s && m->seen==(short)b->s &&
-	 (unsigned short)(doomLevelTime-m->tic)<=(unsigned short)GP_MONSTER_SIGHT_SHARE)
+	 (unsigned short)((unsigned short)doomLevelTime-m->tic)<=
+	 (unsigned short)GP_MONSTER_SIGHT_SHARE)   /* both 16 bits: the subtraction wraps as it must */
 	return m->ans;
     }
 #endif
